strung up the game component. Unsuccessful build with angular cli.
the problem looks like the course I am using as an example is out of date. I will defenitely know how to work with threejs by the end of this project. Basically I need to read the docs to get the most up to date classes, implementaions and practices with threejs.